팝업 컴포넌트 호출 위치 수정(프로젝트 관리 -> 검색 set)
